+++ Semester_1/2022_09_07_Introduction/HW_8/main.c
@@ -1,36 +1,30 @@
 #include <stdio.h>
-#include <stdbool.h>
 #include <stdlib.h>
 
-void swapIntegers(int *a, int *b)
-{
+void swapIntegers(int *a, int *b) {
     *a = (*a ^ *b);
     *b = (*a ^ *b);
     *a = (*a ^ *b);
 }
 
-void reverseSubsegment(int *array, int from, int to) // [from; to)
-{
-    for (int i = 0; i < (to - from) / 2; i++)
-    {
+// [from; to)
+void reverseSubsegment(int *array, size_t from, size_t to) {
+    for (size_t i = 0; i < (to - from) / 2; i++) {
         swapIntegers(&array[from + i], &array[to - i - 1]);
     }
 }
 
-void swapSubsegments(int *array, int prefixSize, int suffixSize)
-{
+void swapSubsegments(int *array, size_t prefixSize, size_t suffixSize) {
     reverseSubsegment(array, 0, prefixSize);
     reverseSubsegment(array, prefixSize, prefixSize + suffixSize);
     reverseSubsegment(array, 0, prefixSize + suffixSize);
 }
 
-int main(void)
-{
+int main(void) {
     printf("Enter prefix size: ");
     int prefixSize = 0;
     scanf("%d", &prefixSize);
-    if (prefixSize < 0)
-    {
+    if (prefixSize < 0) {
         printf("ERROR! Prefix size can't be less than 0!\n");
         return -1;
     }
@@ -38,25 +32,22 @@ int main(void)
     printf("Enter suffix size: ");
     int suffixSize = 0;
     scanf("%d", &suffixSize);
-    if (suffixSize < 0)
-    {
+    if (suffixSize < 0) {
         printf("ERROR! Suffix size can't be less than 0!\n");
         return -1;
     }
     
     int arraySize = prefixSize + suffixSize;
     printf("Enter array (%d elements):\n", arraySize);
-    int *array = malloc(arraySize * sizeof(int));
-    for (int i = 0; i < arraySize; i++)
-    {
+    int *array = malloc((size_t)arraySize * sizeof(int));
+    for (int i = 0; i < arraySize; i++) {
         scanf("%d", &array[i]);
     }
 
-    swapSubsegments(array, prefixSize, suffixSize);
+    swapSubsegments(array, (size_t)prefixSize, (size_t)suffixSize);
 
     printf("Swaped array:\n");
-    for (int i = 0; i < arraySize; i++)
-    {
+    for (int i = 0; i < arraySize; i++) {
         printf("%d ", array[i]);
     }
